Menu changes and Environment changes

+++ SourceCode/Project/src/main/java/models/environments/menus/mainmenu/children/NewGamePage.java
@@ -14,7 +14,7 @@ public class NewGamePage extends AMenu {
         bundle.addPage(new CharacterCreatePage(parentModel));
         AMenuButton button_characterCreate = new AMenuButton(parentModel,500, 200, 200, 50) {
             @Override
-            public boolean onClick(int x, int y) {
+            public boolean onClick(float x, float y) {
                 if(!isInBounds(x, y)) {
                     return false;
                 }
@@ -27,7 +27,7 @@ public class NewGamePage extends AMenu {
 
         AMenuButton button_back = new AMenuButton(parentModel,500, 500, 200, 50) {
             @Override
-            public boolean onClick(int x, int y) {
+            public boolean onClick(float x, float y) {
                 if(!isInBounds(x, y)) {
                     return false;
                 }
